fixed ocr and updated manager

+++ src/controllers/manger.js
@@ -1,10 +1,14 @@
-// make a function that gets show name and date and create it in the database
-const e = require("express");
-const { db } = require("../firebase/firebase");
+const { db, auth } = require("../firebase/firebase");
 // make the document name as the show name and date and put no data in it
 const createShow = async (KenelClubName, date) => {
   const show = await db.collection("shows").doc(`${KenelClubName} - ${date}`);
   try {
+    if ((await show.get()).exists) {
+      return {
+        message: "show already exists",
+        status: 400,
+      };
+    }
     await show.set({});
     return {
       message: "show created",
@@ -12,7 +16,7 @@ const createShow = async (KenelClubName, date) => {
     };
   } catch (error) {
     return {
-      message: "show already exists",
+      error: error,
       status: 400,
     };
   }
@@ -27,6 +31,12 @@ const updateshow = async (KenelClubName, date, newkennelclub, newdate) => {
     .doc(`${newkennelclub} - ${newdate}`);
   //replace the old show with the new one
   try {
+    if (!(await show.get()).exists) {
+      return {
+        message: "show does not exist",
+        status: 400,
+      };
+    }
     const data = await show.get();
     await newshow.set(data.data());
     await show.delete();
@@ -36,7 +46,7 @@ const updateshow = async (KenelClubName, date, newkennelclub, newdate) => {
     };
   } catch (error) {
     return {
-      message: "show does not exist",
+      error: error,
       status: 400,
     };
   }
@@ -46,6 +56,12 @@ const updateshow = async (KenelClubName, date, newkennelclub, newdate) => {
 const deleteShow = async (KenelClubName, date) => {
   const show = await db.collection("shows").doc(`${KenelClubName} - ${date}`);
   try {
+    if (!(await show.get()).exists) {
+      return {
+        message: "show does not exist",
+        status: 400,
+      };
+    }
     await show.delete();
     return {
       message: "show deleted",
@@ -53,10 +69,115 @@ const deleteShow = async (KenelClubName, date) => {
     };
   } catch (error) {
     return {
-      message: "show does not exist",
+      error: error,
       status: 400,
     };
   }
 };
 
-module.exports = { createShow, updateshow, deleteShow };
+const getallshows = async () => {
+  //create an array to store the shows names
+  const shows = [];
+  //get all the shows from the database
+  const showsdata = await db.collection("shows").get();
+  //loop through the shows and push the show name to the array
+  showsdata.forEach((show) => {
+    shows.push(show.id);
+  });
+  //return the array
+  return {
+    shows: shows,
+    message: "shows returned",
+    status: 200,
+  };
+};
+
+const authmanger = async (email, password) => {
+  try {
+    console.log(email, password);
+    //create the user with the email and password
+    const user = await auth.createUser({
+      email: email,
+      password: password,
+    });
+    console.log(user);
+    //return the user
+    return {
+      user: user,
+      message: "manger created",
+      status: 200,
+    };
+  } catch (error) {
+    return {
+      error: error.message,
+      status: 400,
+    };
+  }
+};
+
+const loginmanger = async (email, password) => {
+  try {
+    const user = await auth.signInWithEmailAndPassword({
+      email: email,
+      password: password,
+    });
+    //return the user
+    return {
+      user: user,
+      message: "manger logged in",
+      status: 200,
+    };
+  } catch (error) {
+    return {
+      error: error.message,
+      status: 400,
+    };
+  }
+};
+
+const logoutmanger = async () => {
+  try {
+    await auth.unauth();
+    return {
+      message: "manger logged out",
+      status: 200,
+    };
+  } catch (error) {
+    return {
+      message: "manger not logged in",
+      status: 400,
+    };
+  }
+};
+
+const deleteManger = async (email, password) => {
+  try {
+    console.log(email, password);
+    const user = await auth.signInWithEmailAndPassword({
+      email: email,
+      password: password,
+    });
+    console.log(user);
+    await user.delete();
+    return {
+      message: "manger deleted",
+      status: 200,
+    };
+  } catch (error) {
+    return {
+      error: error.message,
+      status: 400,
+    };
+  }
+};
+
+module.exports = {
+  createShow,
+  updateshow,
+  deleteShow,
+  authmanger,
+  loginmanger,
+  getallshows,
+  logoutmanger,
+  deleteManger,
+};
